moving writeRegistration to registrations
