+++ src/firebase/registrations.js
@@ -1,4 +1,4 @@
-import { db } from './firebase';
+import { db, getStudentDocName } from './firebase';
 
 export async function getAllRegistrations() {
   const querySnapshot = await db.collection("registrations").get();
@@ -7,4 +7,15 @@ export async function getAllRegistrations() {
     registrations.push(doc.data());
   });
   return registrations;
+}
+
+export async function writeRegistration(registration) {
+  const studentDocName = getStudentDocName(registration);
+  return db.collection("registrations").doc(studentDocName).set(registration)
+    .then(()=> {
+      return {success: true};
+    })
+    .catch((error)=> {
+      throw new Error(error);
+    });
 }
